+++ CHANGELOG.md
@@ -3,6 +3,26 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.1.5](https://github.com/ca-cwds/design-system/compare/v1.1.4...v1.1.5) (2019-03-28)
+
+**Note:** Version bump only for package design-system
+
+
+
+
+
+## [1.1.4](https://github.com/ca-cwds/design-system/compare/v1.1.3...v1.1.4) (2019-03-28)
+
+
+### Bug Fixes
+
+* **ci:** fix automate release ([#499](https://github.com/ca-cwds/design-system/issues/499)) ([0c85591](https://github.com/ca-cwds/design-system/commit/0c85591))
+* **docs:** sort Article sidebar links ([#485](https://github.com/ca-cwds/design-system/issues/485)) ([aa0240f](https://github.com/ca-cwds/design-system/commit/aa0240f))
+
+
+
+
+
 ## [1.1.3](https://github.com/ca-cwds/design-system/compare/v1.1.2...v1.1.3) (2019-03-22)